Added default graphic ID and refactored checking for data
E.g. if the weather api changes or I haven't created the graphic data for a weather ID it will deafult to 999 (a plain blue background with no graphic)

Also refactored checking to see if state has the weather data and the loading state. Now exists in app.js rather than in each component.